[feat/#67]: 컴포넌트 분리
